tcg: Infrastructure for managing constant pools
A new shared header tcg-be-ldst-data.h adds new_pool_label,
for registering a tcg_target_ulong to be emitted after the
generated code, plus relocation data to install a pointer
to the data.

A new pointer is added to the TCGContext, so that we dump
the constant pool as data, and not code.

Signed-off-by: Richard Henderson <rth@twiddle.net>