cputlb: Count "partial" and "elided" tlb flushes
Our only statistic so far was "full" tlb flushes, where all mmu_idx
are flushed at the same time.

Now count "partial" tlb flushes where sets of mmu_idx are flushed,
but the set is not maximal.  Account one per mmu_idx flushed, as
that is the unit of work performed.

Also count the number of flushes that we were able to elide because
we can show it was redundant.

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>