tests/tcg/aarch64: Add bti smoke test
This will build with older toolchains, without the upstream support
for -mbranch-protection.  Such a toolchain will produce a warning
in such cases,

ld: warning: /tmp/ccyZt0kq.o: unsupported GNU_PROPERTY_TYPE (5) \
type: 0xc0000000

but the still places the note at the correct location in the binary
for processing by the runtime loader.

But since this warning makes for ugly CI logs, disable it for now.

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>
---
v8: Disable the test for now.