docs/booting.rst: start documenting the boot process
While working on some test cases I realised there was quite a lot of
assumed knowledge about how things boot up. I thought it would be
worth gathering this together in a user facing document where we could
pour in the details and background to the boot process. As it's quite
wordy I thought it should be a separate document to the manual (which
can obviously reference this). So far I've managed almost a thousand
words and haven't actually related anything to QEMU's options yet.

So going forward:

  - is this a useful document to have in docs?
  - are there any other areas to mention?
    - out auto-magic -bios selection seems like something worth covering
  - should we have some worked examples of command lines?
    - I was thinking for example of read-only and pflash examples
  - we should also describe why direct kernel boots exits
    - and also the fact they are not that direct (some code executes
      before a kernel - even if it's less than a full firmware)

Submitted for comment....

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>