tcg/i386: Require BYTE_REGS for INDEX_op_qemu_st_i32 bytes
This is the payoff for the target_lookup_constraint conversion.

Since we can now examine the arguments to INDEX_op_qemu_st_i32,
we can tell whether or not we are storing a byte, and return a
different set of constraints.  Because this only matters for the
32-bit host, we do not need to care about INDEX_op_qemu_st_i64,
since that will only be used for storing 64-bit values.

We now require no reserved registers for user-only memory ops.

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>